major command reorganization
