+++ build.gradle
@@ -83,6 +83,7 @@ dependencies {
     testImplementation 'junit:junit:4.12'
 }
 
+
 // Simulation configuration (e.g. environment variables).
 wpi.sim.addGui().defaultEnabled = true
 wpi.sim.addDriverstation()
