+++ src/main/java/frc/robot/RobotContainer.java
@@ -16,19 +16,19 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
-import frc.robot.commands.AutoDoNothing;
-import frc.robot.commands.AutoShootLeaveTarmac;
-import frc.robot.commands.AutoTwoBall;
-import frc.robot.commands.CollectGroup;
-import frc.robot.commands.CommandHoodTuning;
-import frc.robot.commands.DriveFieldRelative;
-import frc.robot.commands.JoystickClimb;
-import frc.robot.commands.LeaveTarmack;
-import frc.robot.commands.LimelightAlign;
-import frc.robot.commands.MoveLongArms;
-import frc.robot.commands.MoveShortArms;
-import frc.robot.commands.MoveTower;
-import frc.robot.commands.PrimitiveShooterTuning;
+import frc.robot.commands.auto.AutoDoNothing;
+import frc.robot.commands.auto.AutoShootLeaveTarmac;
+import frc.robot.commands.auto.AutoTwoBall;
+import frc.robot.commands.auto.LeaveTarmack;
+import frc.robot.commands.base.DriveFieldRelative;
+import frc.robot.commands.base.LimelightAlign;
+import frc.robot.commands.climber.JoystickClimb;
+import frc.robot.commands.climber.MoveLongArms;
+import frc.robot.commands.climber.MoveShortArms;
+import frc.robot.commands.collector.CollectGroup;
+import frc.robot.commands.hood.CommandHoodTuning;
+import frc.robot.commands.shooter.PrimitiveShooterTuning;
+import frc.robot.commands.tower.MoveTower;
 import frc.robot.subsystems.Base;
 import frc.robot.subsystems.Climbers;
 import frc.robot.subsystems.Collector;
