+++ src/main/java/frc/robot/commands/auto/LeaveTarmack.java
@@ -1,0 +1,26 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package frc.robot.commands.auto;
+
+import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import frc.robot.commands.base.Navigate;
+import frc.robot.subsystems.Base;
+
+// NOTE:  Consider using this command inline, rather than writing a subclass.  For more
+// information, see:
+// https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
+public class LeaveTarmack extends SequentialCommandGroup {
+
+  public static double TARMAC_WIDTH = 93;
+  public static double ROBOT_WIDTH = 24;
+  public static double TRAVEL_DISTANCE = TARMAC_WIDTH + ROBOT_WIDTH;
+
+  /** Creates a new LeaveTarmack. */
+  public LeaveTarmack(Base base) {
+    // Add your commands in the addCommands() call, e.g.
+    // addCommands(new FooCommand(), new BarCommand());
+    addCommands(new Navigate(base, TRAVEL_DISTANCE));
+  }
+}
