+++ src/main/java/frc/robot/subsystems/Dashboard.java
@@ -10,7 +10,7 @@ import frc.robot.BaseConstants;
 import frc.robot.ClimberModule;
 import frc.robot.HoodConstants;
 import frc.robot.ShooterConstants;
-import frc.robot.commands.LimelightAlign;
+import frc.robot.commands.base.LimelightAlign;
 
 public class Dashboard extends SubsystemBase {
 
@@ -69,112 +69,94 @@ public class Dashboard extends SubsystemBase {
 
   /** Creates a new Dashboard. */
   public Dashboard(
-    Base base,
-    Collector collector,
-    Shooter shooter,
-    Climbers climbers
-  ) {
+      Base base,
+      Collector collector,
+      Shooter shooter,
+      Climbers climbers) {
     this.base = base;
     this.collector = collector;
     this.shooter = shooter;
     this.climbers = climbers;
   }
 
-  public void initialize() {}
+  public void initialize() {
+  }
 
   public void publishInitialDashboard() {
-    //public information about subsystems
-    //this info is read only and put on dashboard for display purposes only
+    // public information about subsystems
+    // this info is read only and put on dashboard for display purposes only
     if (base != null) {
       SmartDashboard.putNumber(DASH_BASE_YAW, this.base.getSensorYaw());
       SmartDashboard.putNumber(
-        DASH_BASE_FLPOS,
-        this.base.frontLeftModule.getDrivePosition()
-      );
+          DASH_BASE_FLPOS,
+          this.base.frontLeftModule.getDrivePosition());
       SmartDashboard.putNumber(
-        DASH_BASE_FRPOS,
-        this.base.frontRightModule.getDrivePosition()
-      );
+          DASH_BASE_FRPOS,
+          this.base.frontRightModule.getDrivePosition());
       SmartDashboard.putNumber(
-        DASH_BASE_BLPOS,
-        this.base.backLeftModule.getDrivePosition()
-      );
+          DASH_BASE_BLPOS,
+          this.base.backLeftModule.getDrivePosition());
       SmartDashboard.putNumber(
-        DASH_BASE_BRPOS,
-        this.base.backRightModule.getDrivePosition()
-      );
-    } else if (collector != null) {} else if (shooter != null) {
+          DASH_BASE_BRPOS,
+          this.base.backRightModule.getDrivePosition());
+    } else if (collector != null) {
+    } else if (shooter != null) {
       SmartDashboard.putNumber(
-        Dashboard.DASH_SHOOTER_VELOCITY,
-        this.shooter.getMotorOutputPercent()
-      );
+          Dashboard.DASH_SHOOTER_VELOCITY,
+          this.shooter.getMotorOutputPercent());
       SmartDashboard.putNumber(
-        Dashboard.DASH_SHOOTER_RPMS,
-        this.shooter.getMotorRpms()
-      );
+          Dashboard.DASH_SHOOTER_RPMS,
+          this.shooter.getMotorRpms());
     } else if (climbers != null) {
       SmartDashboard.putNumber(
-        Dashboard.DASH_CLIMBER_LONG_ARM_POSITION,
-        this.climbers.longClimberModule.getPosition()
-      );
+          Dashboard.DASH_CLIMBER_LONG_ARM_POSITION,
+          this.climbers.longClimberModule.getPosition());
       SmartDashboard.putNumber(
-        Dashboard.DASH_CLIMBER_SHORT_ARM_POSITION,
-        this.climbers.shortClimberModule.getPosition()
-      );
+          Dashboard.DASH_CLIMBER_SHORT_ARM_POSITION,
+          this.climbers.shortClimberModule.getPosition());
     }
 
     SmartDashboard.putNumber(Dashboard.DASH_HOOD_P, HoodConstants.kGains.kP);
     SmartDashboard.putNumber(Dashboard.DASH_HOOD_I, HoodConstants.kGains.kI);
     SmartDashboard.putNumber(Dashboard.DASH_HOOD_D, HoodConstants.kGains.kD);
     SmartDashboard.putNumber(
-      Dashboard.DASH_TARGET_TRACKER_KP,
-      BaseConstants.targetTrackerGains.kP
-    );
+        Dashboard.DASH_TARGET_TRACKER_KP,
+        BaseConstants.targetTrackerGains.kP);
     SmartDashboard.putNumber(
-      Dashboard.DASH_TARGET_TRACKER_KI,
-      BaseConstants.targetTrackerGains.kI
-    );
+        Dashboard.DASH_TARGET_TRACKER_KI,
+        BaseConstants.targetTrackerGains.kI);
     SmartDashboard.putNumber(
-      Dashboard.DASH_TARGET_TRACKER_KD,
-      BaseConstants.targetTrackerGains.kD
-    );
+        Dashboard.DASH_TARGET_TRACKER_KD,
+        BaseConstants.targetTrackerGains.kD);
 
     SmartDashboard.putNumber(
-      Dashboard.DASH_HOOD_TICKS,
-      HoodConstants.IDLE_TICK_COUNT
-    );
+        Dashboard.DASH_HOOD_TICKS,
+        HoodConstants.IDLE_TICK_COUNT);
     SmartDashboard.putNumber(Dashboard.DASH_HOOD_OUTPUT, 0);
     SmartDashboard.putNumber(Dashboard.DASH_HOOD_DRAW, 0);
     SmartDashboard.putNumber(
-      Dashboard.DASH_CLIMBER_LIMITER,
-      ClimberModule.CLIMBER_LIMITER
-    );
+        Dashboard.DASH_CLIMBER_LIMITER,
+        ClimberModule.CLIMBER_LIMITER);
     SmartDashboard.putNumber(
-      Dashboard.BASE_ALIGN_LIMIT,
-      LimelightAlign.TRACKER_LIMIT_DEFAULT
-    );
+        Dashboard.BASE_ALIGN_LIMIT,
+        LimelightAlign.TRACKER_LIMIT_DEFAULT);
     SmartDashboard.putNumber(
-      DASH_SHOOTER_RPMS,
-      ShooterConstants.SHOOTER_POWER_HUB_HIGH
-    );
+        DASH_SHOOTER_RPMS,
+        ShooterConstants.SHOOTER_POWER_HUB_HIGH);
 
     SmartDashboard.putNumber(
-      Dashboard.DASH_SHOOTER_TARGET_RPMS,
-      ShooterConstants.SHOOTER_POWER_HUB_HIGH
-    );
+        Dashboard.DASH_SHOOTER_TARGET_RPMS,
+        ShooterConstants.SHOOTER_POWER_HUB_HIGH);
     SmartDashboard.putBoolean(Dashboard.DASH_SHOOTER_READY, false);
     SmartDashboard.putNumber(
-      Dashboard.DASH_SHOOTER_P,
-      ShooterConstants.kGains.kP
-    );
+        Dashboard.DASH_SHOOTER_P,
+        ShooterConstants.kGains.kP);
     SmartDashboard.putNumber(
-      Dashboard.DASH_SHOOTER_FF,
-      ShooterConstants.kGains.kF
-    );
+        Dashboard.DASH_SHOOTER_FF,
+        ShooterConstants.kGains.kF);
     SmartDashboard.putNumber(
-      DASH_LIMELIGHT_ANGLE_OFFSET,
-      LimelightAlign.TRACKER_OFFSET
-    );
+        DASH_LIMELIGHT_ANGLE_OFFSET,
+        LimelightAlign.TRACKER_OFFSET);
   }
 
   @Override
